move implementations to impls.rs
